feat: replace confusion_matrix & printer
